feat: Extracting ErrorText into an atom
