+++ src/components/Atoms/ErrorText/ErrorText.md
@@ -1,0 +1,5 @@
+# ErrorText
+
+```jsx
+<ErrorText size="sm">This is an error</ErrorText>
+```
