+++ src/components/Atoms/Input/Input.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import styled from 'styled-components';
 
 import Text from '../Text/Text';
-import alertIcon from './assets/error-alert-icon.png';
+import ErrorText from '../ErrorText/ErrorText';
 import hideVisually from '../../../theme/shared/hideVisually';
 import spacing from '../../../theme/shared/spacing';
 
@@ -49,28 +49,6 @@ const Label = styled.label`
   flex-direction: column;
   font-weight: bold;
   color: ${({ theme }) => theme.color('grey')};
-`;
-
-/**
- * Text error component
- */
-const ErrorText = styled(Text)`
-  display: inline-block;
-  color: red;
-  font-weight: ${({ weight }) => weight};
-  margin-top: ${spacing('sm')};
-  :before {
-    position: relative;
-    content: '';
-    top: 2px;
-    margin-right: 6px;
-    background: url(${alertIcon}) left 0/100% no-repeat;
-    width: 18px;
-    height: 19px;
-    display: inline-block;
-    vertical-align: top;
-    color: ${({ theme }) => theme.color('red')};
-  }
 `;
 
 /**
